chore: refactor auth to authn server pkgs
