+++ internal/server/authn/method/oidc/server_test.go
@@ -21,8 +21,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.flipt.io/flipt/internal/config"
-	authoidc "go.flipt.io/flipt/internal/server/auth/method/oidc"
-	oidctesting "go.flipt.io/flipt/internal/server/auth/method/oidc/testing"
+	authoidc "go.flipt.io/flipt/internal/server/authn/method/oidc"
+	oidctesting "go.flipt.io/flipt/internal/server/authn/method/oidc/testing"
 	"go.flipt.io/flipt/rpc/flipt/auth"
 	"go.uber.org/zap/zaptest"
 	"golang.org/x/net/html"
