+++ internal/server/middleware/grpc/middleware_test.go
@@ -11,8 +11,8 @@ import (
 	"go.flipt.io/flipt/internal/common"
 	"go.flipt.io/flipt/internal/config"
 	"go.flipt.io/flipt/internal/server"
-	"go.flipt.io/flipt/internal/server/auth/method/token"
-	authmiddlewaregrpc "go.flipt.io/flipt/internal/server/auth/middleware/grpc"
+	"go.flipt.io/flipt/internal/server/authn/method/token"
+	authmiddlewaregrpc "go.flipt.io/flipt/internal/server/authn/middleware/grpc"
 	servereval "go.flipt.io/flipt/internal/server/evaluation"
 	"go.flipt.io/flipt/internal/storage"
 	flipt "go.flipt.io/flipt/rpc/flipt"
